tp/requirement.txt
awscli==1.16.230
backports.shutil-get-terminal-size==1.0.0
boto3==1.9.220
botocore==1.12.220
colorama==0.3.9
decorator==4.4.0
Django==1.11.23
django-google-maps==0.11.0
djangorestframework==3.9.4
docutils==0.15.2
enum34==1.1.6
futures==3.3.0
ipdb==0.12.2
ipython==5.8.0
ipython-genutils==0.2.0
jmespath==0.9.4
model-mommy==1.6.0
mysqlclient==1.4.3
pathlib2==2.3.4
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==1.0.16
ptyprocess==0.6.0
pyasn1==0.4.7
Pygments==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
rsa==3.4.2
s3transfer==0.2.1
scandir==1.10.0
simplegeneric==0.8.1
six==1.12.0
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
